moved location_geodata to extapi
